Switch Code Revision
Revised Switch script so that it records the default state for connected objects and reverses them when off (rather than strictly setting objects off when switch is on and vice versa).
So now we can have connected objects that are off when switch is off but also become active when switch is active.
Revised visual indication so rather than connected objects literally become an inactive object, we just turn off the collider and make the object transparent (0.2f alpha).