NeedleOnTriggerEnter Addition
Originally, this script actually had implementation for TriggerStay but nothing for TriggerEnter.
I realize the lack of implementation for TriggerEnter is causing a "delayed" reaction to stopping needle for OnTriggers with ground/walls since Enter is usually called before Stay. Hence, I copied TriggerStay's implementation to TriggerEnter.
From testing, it looks like if the needle already overlaps with a ground/wall and then with thrown, it will stop much faster (still takes a fraction of a second but at least it won't go out as far before stopping late).