NeedleRotation CircleOffset
When equipped, the needle rotates around player based on the mouse aim.
For the needle, there's a certain amount of circle Offset from the player's center that it rotates around.
Before, this circle Offset was a fixed 0.9f. I added the option to make that circle offset a public value (0.9f by default) so we can play around with it. This circle offset impacts where needle starts when it's thrown.